Generate random, unique card in controller
This code generates a random card for a cycle in a given round. The logic will grab a random card for a deck in the first cycle, and in subsequent cycles generate a random card from the cards that were answered incorrectly in the previous cycle. There are still some problems with the code which are noted in the comments of the cards_controller.rb